Draw the box cursor inside the character cell
Add local VTE modification to draw the box cursor (which is used instead
of the block cursor when the window does not have the input focus)
completely inside the character cell.

By default VTE draws the box cursor outside the character cell, which
was probably done to avoid covering up the character under the cursor;
however, the box outside the character cell may not fit in the window if
the terminal is configured to have no padding, and in this case the box
cursor with missing sides looks ugly.  Change the code to draw the box
cursor completely inside the character cell, like all other terminal
emulators do.