Use VGA-like colors when legacy background colors are used
Add local VTE modification to use VGA-like colors when legacy background
colors are used.  This should make programs which assume VGA-like colors
usable out of the box (e.g., some programs which output light gray or
cyan text on the blue background will look as intended even when those
colors in the custom palette are modified to be readable over the
default background), but does not interfere with using a custom palette
for foreground colors over the default background.

The VGA-like colors are technically configurable using the existing
`vte_terminal_set_colors()` API, but the numbers used for those colors
are unofficial and may change.  In addition, those colors are at the end
of the palette, and the API does not provide a way to set just those
colors, therefore setting them may be very inconvenient, because a
completely filled palette with 256 terminal colors and some more
internally used colors must be provided.